// src/player/dto/player-dto.ts
export type PlayerPageDTO = {
    name: string
    project_id: string
    rpg: string
    password: string
}

export type PlayerAcess = {
    name: string
    email: string
    project_id: string
    password: string
}

export type PlayerCardDTO = {
    name: string
    email: string
    project_id: string
    password: string
}